Rewrite the ByteArrayInputStream for a major performance improvement, and fix the test code that used it so remove some warnings with type conversions.
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk/activemq-cpp@928770 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/activemq/wireformat/openwire/marshal/v1/ActiveMQObjectMessageMarshallerTest.cpp
@@ -100,7 +100,7 @@ void ActiveMQObjectMessageMarshallerTest::testLooseMarshal() {
         marshaller.looseMarshal( &openWireFormat, &outCommand, &dataOut );
 
         // Now read it back in and make sure it's all right.
-        ByteArrayInputStream bais( baos.toByteArray(), baos.size() );
+        ByteArrayInputStream bais( baos.toByteArray(), (int)baos.size() );
         DataInputStream dataIn( &bais );
         unsigned char dataType = dataIn.readByte();
         CPPUNIT_ASSERT( dataType == outCommand.getDataStructureType() );
@@ -158,7 +158,7 @@ void ActiveMQObjectMessageMarshallerTest::testTightMarshal() {
         marshaller.tightMarshal2( &openWireFormat, &outCommand, &dataOut, &bs );
 
         // Now read it back in and make sure it's all right.
-        ByteArrayInputStream bais( baos.toByteArray(), baos.size() );
+        ByteArrayInputStream bais( baos.toByteArray(), (int)baos.size() );
         DataInputStream dataIn( &bais );
 
         unsigned char dataType = dataIn.readByte();
