+++ src/test/decaf/io/FilterOutputStreamTest.cpp
@@ -77,7 +77,7 @@ void FilterOutputStreamTest::testWrite1() {
         ByteArrayOutputStream baos;
         FilterOutputStream os( &baos );
         os.write( (unsigned char*)&testString[0], (int)testString.size(), 0, (int)testString.size() );
-        ByteArrayInputStream bais( baos.toByteArray(), baos.size() );
+        ByteArrayInputStream bais( baos.toByteArray(), (int)baos.size() );
         os.flush();
         CPPUNIT_ASSERT_MESSAGE( "Bytes not written after flush",
                                 bais.available() == (int)testString.length() );
@@ -99,7 +99,7 @@ void FilterOutputStreamTest::testWrite2() {
         ByteArrayOutputStream baos;
         FilterOutputStream os( &baos );
         os.write('t');
-        ByteArrayInputStream bais( baos.toByteArray(), baos.size() );
+        ByteArrayInputStream bais( baos.toByteArray(), (int)baos.size() );
         os.flush();
         CPPUNIT_ASSERT_MESSAGE( "Byte not written after flush", 1 == bais.available() );
         unsigned char wbytes[1];
